Refactor api Add comments for exported functions
