+++ api/errors.go
@@ -11,6 +11,8 @@ func (err *GoHabitsError) Error() string {
 	return err.msg
 }
 
+// NewGoHabitsError is constructor for GoHabitsError that includes
+// program status code, uri/path, and a simple message string.
 func NewGoHabitsError(message string, code int, path string) *GoHabitsError {
 	return &GoHabitsError{message, code, path}
 }
